requirements.txt
click
graphviz
prov
rdflib-jsonld
pyld
pydot
requests
pyyaml
rdflib
boutiques
beautifulsoup4
deepdiff
pybids
rfc3986